+++ NEWS.md
@@ -1,3 +1,6 @@
 # conflicted 0.1.0.9000
 
 * Added a `NEWS.md` file to track changes to the package.
+
+* conflicted now listens for `detach()` events and removes conflicts that
+  are removed by detaching a package (#5)